Add F7Page behavior and use it instead of View sub classes
