+++ src/app/scripts/views/tab2/index.js
@@ -1,7 +1,16 @@
-var MainItemView = require('../mainItemView');
+var Marionette = require('marionette');
+var F7Page = require('../f7page');
 
-module.exports = MainItemView.extend({
+module.exports = Marionette.View.extend({
     template: require('../../templates/tab2/index.hbs'),
+
+    behaviors: [
+      {
+        behaviorClass: F7Page,
+        isRoot: true
+      }
+    ],
+
     attributes: function() {
         return {
             id: "tab2",
